+++ scm-ui/eslint-config/src/__resources__/AllowRootImport.tsx
@@ -21,36 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-module.exports = {
-  extends: [
-    "react-app",
-    "plugin:prettier/recommended",
-    "plugin:flowtype/recommended"
-  ],
-  rules: {
-    semi: ["error", "always"],
-    quotes: ["error", "double"],
-    "jsx-a11y/href-no-hash": [0],
-    "flowtype/no-types-missing-file-annotation": 2,
-    "no-console": "error"
-  },
-  overrides: [
-    {
-      files: ["*.ts", "*.tsx"],
-      parser: "@typescript-eslint/parser",
-      extends: [
-        "react-app",
-        "plugin:prettier/recommended",
-        "plugin:@typescript-eslint/recommended"
-      ],
-      rules: {
-        semi: ["error", "always"],
-        quotes: ["error", "double"],
-        "jsx-a11y/href-no-hash": [0],
-        "@typescript-eslint/explicit-function-return-type": "off",
-        "@typescript-eslint/ban-ts-ignore": "warn",
-        "no-console": "error"
-      }
-    }
-  ]
-};
+
+import React, { FC } from "react";
+import { Button } from "@scm-manager/ui-components";
+
+const SpecialButton: FC = () => <Button color="primary">Special</Button>;
+
+export default SpecialButton;